requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
python-dotenv
requests
google-auth